Added module docstrings/infos and renamed api_client to client
